+++ cmd/hidden_lake/helpers/distributor/_daemon/restart_hld.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# root mode
+systemctl restart hidden_lake_distributor.service