+++ cmd/hidden_lake/helpers/distributor/pkg/app/errors.go
@@ -1,0 +1,19 @@
+package app
+
+const (
+	errPrefix = "cmd/hidden_lake/helpers/loader/pkg/app = "
+)
+
+type SAppError struct {
+	str string
+}
+
+func (err *SAppError) Error() string {
+	return errPrefix + err.str
+}
+
+var (
+	ErrRunning = &SAppError{"app running"}
+	ErrService = &SAppError{"service"}
+	ErrClose   = &SAppError{"close"}
+)