+++ cmd/hidden_lake/helpers/distributor/pkg/client/client_test.go
@@ -1,0 +1,16 @@
+package client
+
+import "testing"
+
+func TestError(t *testing.T) {
+	t.Parallel()
+
+	str := "value"
+	err := &SClientError{str}
+	if err.Error() != errPrefix+str {
+		t.Error("incorrect err.Error()")
+		return
+	}
+}
+
+func TestNothing(_ *testing.T) {}