+++ cmd/hidden_lake/helpers/distributor/pkg/config/types.go
@@ -1,0 +1,11 @@
+package config
+
+import "github.com/number571/go-peer/cmd/hidden_lake/helpers/distributor/internal/config"
+
+type IConfigSettings interface {
+	config.IConfigSettings
+}
+
+type SConfigSettings struct {
+	config.SConfigSettings
+}