+++ cmd/hidden_lake/helpers/distributor/pkg/settings/settings.go
@@ -1,0 +1,20 @@
+package settings
+
+const (
+	CServiceName     = "HLD"
+	CServiceFullName = "hidden-lake-distributor"
+)
+
+const (
+	CPathYML = "hld.yml"
+)
+
+const (
+	CDefaultHTTPAddress = "127.0.0.1:9521"
+)
+
+const (
+	CHandleIndexPath             = "/api/index"
+	CHandleConfigSettings        = "/api/config/settings"
+	CHandleNetworkDistributePath = "/api/network/distribute"
+)